% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{AuditConfig}
\alias{AuditConfig}
\title{AuditConfig Object}
\usage{
AuditConfig(service = NULL, auditLogConfigs = NULL,
  exemptedMembers = NULL)
}
\arguments{
\item{service}{Specifies a service that will be enabled for audit logging}

\item{auditLogConfigs}{The configuration for logging of each type of permission}

\item{exemptedMembers}{Specifies the identities that are exempted from 'data access' audit}
}
\value{
AuditConfig object
}
\description{
AuditConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the audit configuration for a service.It consists of which permission types are logged, and what identities, ifany, are exempted from logging.An AuditConifg must have one or more AuditLogConfigs.
}
