% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{BeginTransactionRequest}
\alias{BeginTransactionRequest}
\title{BeginTransactionRequest Object}
\usage{
BeginTransactionRequest(options = NULL)
}
\arguments{
\item{options}{Required}
}
\value{
BeginTransactionRequest object
}
\description{
BeginTransactionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for BeginTransaction.
}
