% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ChildLink}
\alias{ChildLink}
\title{ChildLink Object}
\usage{
ChildLink(type = NULL, childIndex = NULL, variable = NULL)
}
\arguments{
\item{type}{The type of the link}

\item{childIndex}{The node to which the link points}

\item{variable}{Only present if the child node is SCALAR and corresponds}
}
\value{
ChildLink object
}
\description{
ChildLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata associated with a parent-child relationship appearing in aPlanNode.
}
