% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CloudAuditOptions}
\alias{CloudAuditOptions}
\title{CloudAuditOptions Object}
\usage{
CloudAuditOptions()
}
\value{
CloudAuditOptions object
}
\description{
CloudAuditOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Write a Cloud Audit log
}
