% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CommitRequest}
\alias{CommitRequest}
\title{CommitRequest Object}
\usage{
CommitRequest(singleUseTransaction = NULL, mutations = NULL,
  transactionId = NULL)
}
\arguments{
\item{singleUseTransaction}{Execute mutations in a temporary transaction}

\item{mutations}{The mutations to be executed when this transaction commits}

\item{transactionId}{Commit a previously-started transaction}
}
\value{
CommitRequest object
}
\description{
CommitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Commit.
}
