% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CommitResponse}
\alias{CommitResponse}
\title{CommitResponse Object}
\usage{
CommitResponse(commitTimestamp = NULL)
}
\arguments{
\item{commitTimestamp}{The Cloud Spanner timestamp at which the transaction committed}
}
\value{
CommitResponse object
}
\description{
CommitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Commit.
}
