% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CreateDatabaseMetadata}
\alias{CreateDatabaseMetadata}
\title{CreateDatabaseMetadata Object}
\usage{
CreateDatabaseMetadata(database = NULL)
}
\arguments{
\item{database}{The database being created}
}
\value{
CreateDatabaseMetadata object
}
\description{
CreateDatabaseMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata type for the operation returned byCreateDatabase.
}
