% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CreateDatabaseRequest}
\alias{CreateDatabaseRequest}
\title{CreateDatabaseRequest Object}
\usage{
CreateDatabaseRequest(createStatement = NULL, extraStatements = NULL)
}
\arguments{
\item{createStatement}{Required}

\item{extraStatements}{An optional list of DDL statements to run inside the newly created}
}
\value{
CreateDatabaseRequest object
}
\description{
CreateDatabaseRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for CreateDatabase.
}
