% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CreateInstanceMetadata}
\alias{CreateInstanceMetadata}
\title{CreateInstanceMetadata Object}
\usage{
CreateInstanceMetadata(instance = NULL, startTime = NULL,
  cancelTime = NULL, endTime = NULL)
}
\arguments{
\item{instance}{The instance being created}

\item{startTime}{The time at which the}

\item{cancelTime}{The time at which this operation was cancelled}

\item{endTime}{The time at which this operation failed or was completed successfully}
}
\value{
CreateInstanceMetadata object
}
\description{
CreateInstanceMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata type for the operation returned byCreateInstance.
}
