% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{CreateInstanceRequest}
\alias{CreateInstanceRequest}
\title{CreateInstanceRequest Object}
\usage{
CreateInstanceRequest(instanceId = NULL, instance = NULL)
}
\arguments{
\item{instanceId}{Required}

\item{instance}{Required}
}
\value{
CreateInstanceRequest object
}
\description{
CreateInstanceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for CreateInstance.
}
\seealso{
Other CreateInstanceRequest functions: \code{\link{projects.instances.create}}
}
