% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Database}
\alias{Database}
\title{Database Object}
\usage{
Database(state = NULL, name = NULL)
}
\arguments{
\item{state}{Output only}

\item{name}{Required}
}
\value{
Database object
}
\description{
Database Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Spanner database.
}
