% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Delete}
\alias{Delete}
\title{Delete Object}
\usage{
Delete(table = NULL, keySet = NULL)
}
\arguments{
\item{table}{Required}

\item{keySet}{Required}
}
\value{
Delete object
}
\description{
Delete Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Arguments to delete operations.
}
