% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ExecuteSqlRequest.paramTypes}
\alias{ExecuteSqlRequest.paramTypes}
\title{ExecuteSqlRequest.paramTypes Object}
\usage{
ExecuteSqlRequest.paramTypes()
}
\value{
ExecuteSqlRequest.paramTypes object
}
\description{
ExecuteSqlRequest.paramTypes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
It is not always possible for Cloud Spanner to infer the right SQL typefrom a JSON value.  For example, values of type `BYTES` and valuesof type `STRING` both appear in params as JSON strings.In these cases, `param_types` can be used to specify the exactSQL type for some or all of the SQL query parameters. See thedefinition of Type for more informationabout SQL types.
}
\seealso{
Other ExecuteSqlRequest functions: \code{\link{ExecuteSqlRequest.params}},
  \code{\link{ExecuteSqlRequest}}
}
