% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ExecuteSqlRequest}
\alias{ExecuteSqlRequest}
\title{ExecuteSqlRequest Object}
\usage{
ExecuteSqlRequest(ExecuteSqlRequest.paramTypes = NULL,
  ExecuteSqlRequest.params = NULL, transaction = NULL, resumeToken = NULL,
  paramTypes = NULL, sql = NULL, params = NULL, queryMode = NULL)
}
\arguments{
\item{ExecuteSqlRequest.paramTypes}{The \link{ExecuteSqlRequest.paramTypes} object or list of objects}

\item{ExecuteSqlRequest.params}{The \link{ExecuteSqlRequest.params} object or list of objects}

\item{transaction}{The transaction to use}

\item{resumeToken}{If this request is resuming a previously interrupted SQL query}

\item{paramTypes}{It is not always possible for Cloud Spanner to infer the right SQL type}

\item{sql}{Required}

\item{params}{The SQL query string can contain parameter placeholders}

\item{queryMode}{Used to control the amount of debugging information returned in}
}
\value{
ExecuteSqlRequest object
}
\description{
ExecuteSqlRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for ExecuteSql andExecuteStreamingSql.
}
\seealso{
Other ExecuteSqlRequest functions: \code{\link{ExecuteSqlRequest.paramTypes}},
  \code{\link{ExecuteSqlRequest.params}}
}
