% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Field}
\alias{Field}
\title{Field Object}
\usage{
Field(type = NULL, name = NULL)
}
\arguments{
\item{type}{The type of the field}

\item{name}{The name of the field}
}
\value{
Field object
}
\description{
Field Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message representing a single field of a struct.
}
