% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{GetDatabaseDdlResponse}
\alias{GetDatabaseDdlResponse}
\title{GetDatabaseDdlResponse Object}
\usage{
GetDatabaseDdlResponse(statements = NULL)
}
\arguments{
\item{statements}{A list of formatted DDL statements defining the schema of the database}
}
\value{
GetDatabaseDdlResponse object
}
\description{
GetDatabaseDdlResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for GetDatabaseDdl.
}
