% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Instance.labels}
\alias{Instance.labels}
\title{Instance.labels Object}
\usage{
Instance.labels()
}
\value{
Instance.labels object
}
\description{
Instance.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Labels are a flexible and lightweight mechanism for organizing cloudresources into groups that reflect a customer's organizational needs anddeployment strategies. Cloud Labels can be used to filter collections ofresources. They can be used to control how resource metrics are aggregated.And they can be used as arguments to policy management rules (e.g. route,firewall, load balancing, etc.). * Label keys must be between 1 and 63 characters long and must conform to   the following regular expression: `[a-z]([-a-z0-9]*[a-z0-9])?`. * Label values must be between 0 and 63 characters long and must conform   to the regular expression `([a-z]([-a-z0-9]*[a-z0-9])?)?`. * No more than 64 labels can be associated with a given resource.See https://goo.gl/xmQnxf for more information on and examples of labels.If you plan to use labels in your own code, please note that additionalcharacters may be allowed in the future. And so you are advised to use aninternal label representation, such as JSON, which doesn't rely uponspecific characters being disallowed.  For example, representing labelsas the string:  name + '_' + value  would prove problematic if we were toallow '_' in a future release.
}
\seealso{
Other Instance functions: \code{\link{Instance}}
}
