% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{InstanceConfig}
\alias{InstanceConfig}
\title{InstanceConfig Object}
\usage{
InstanceConfig(name = NULL, displayName = NULL)
}
\arguments{
\item{name}{A unique identifier for the instance configuration}

\item{displayName}{The name of this instance configuration as it appears in UIs}
}
\value{
InstanceConfig object
}
\description{
InstanceConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A possible configuration for a Cloud Spanner instance. Configurationsdefine the geographic placement of nodes and their replication.
}
