% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{KeyRange}
\alias{KeyRange}
\title{KeyRange Object}
\usage{
KeyRange(startClosed = NULL, startOpen = NULL, endOpen = NULL,
  endClosed = NULL)
}
\arguments{
\item{startClosed}{If the start is closed, then the range includes all rows whose}

\item{startOpen}{If the start is open, then the range excludes rows whose first}

\item{endOpen}{If the end is open, then the range excludes rows whose first}

\item{endClosed}{If the end is closed, then the range includes all rows whose}
}
\value{
KeyRange object
}
\description{
KeyRange Object
}
