% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ListDatabasesResponse}
\alias{ListDatabasesResponse}
\title{ListDatabasesResponse Object}
\usage{
ListDatabasesResponse(nextPageToken = NULL, databases = NULL)
}
\arguments{
\item{nextPageToken}{`next_page_token` can be sent in a subsequent}

\item{databases}{Databases that matched the request}
}
\value{
ListDatabasesResponse object
}
\description{
ListDatabasesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for ListDatabases.
}
