% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ListInstanceConfigsResponse}
\alias{ListInstanceConfigsResponse}
\title{ListInstanceConfigsResponse Object}
\usage{
ListInstanceConfigsResponse(nextPageToken = NULL, instanceConfigs = NULL)
}
\arguments{
\item{nextPageToken}{`next_page_token` can be sent in a subsequent}

\item{instanceConfigs}{The list of requested instance configurations}
}
\value{
ListInstanceConfigsResponse object
}
\description{
ListInstanceConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for ListInstanceConfigs.
}
