% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ListInstancesResponse}
\alias{ListInstancesResponse}
\title{ListInstancesResponse Object}
\usage{
ListInstancesResponse(instances = NULL, nextPageToken = NULL)
}
\arguments{
\item{instances}{The list of requested instances}

\item{nextPageToken}{`next_page_token` can be sent in a subsequent}
}
\value{
ListInstancesResponse object
}
\description{
ListInstancesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for ListInstances.
}
