% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Mutation}
\alias{Mutation}
\title{Mutation Object}
\usage{
Mutation(delete = NULL, insert = NULL, insertOrUpdate = NULL,
  update = NULL, replace = NULL)
}
\arguments{
\item{delete}{Delete rows from a table}

\item{insert}{Insert new rows in a table}

\item{insertOrUpdate}{Like insert, except that if the row already exists, then}

\item{update}{Update existing rows in a table}

\item{replace}{Like insert, except that if the row already exists, it is}
}
\value{
Mutation object
}
\description{
Mutation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A modification to one or more Cloud Spanner rows.  Mutations can beapplied to a Cloud Spanner database by sending them in aCommit call.
}
