% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{PartialResultSet}
\alias{PartialResultSet}
\title{PartialResultSet Object}
\usage{
PartialResultSet(stats = NULL, chunkedValue = NULL, metadata = NULL,
  values = NULL, resumeToken = NULL)
}
\arguments{
\item{stats}{Query plan and execution statistics for the query that produced this}

\item{chunkedValue}{If true, then the final value in values is chunked, and must}

\item{metadata}{Metadata about the result set, such as row type information}

\item{values}{A streamed result set consists of a stream of values, which might}

\item{resumeToken}{Streaming calls might be interrupted for a variety of reasons, such}
}
\value{
PartialResultSet object
}
\description{
PartialResultSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Partial results from a streaming read or SQL query. Streaming reads andSQL queries better tolerate large result sets, large rows, and largevalues, but are a little trickier to consume.
}
