% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{PlanNode.executionStats}
\alias{PlanNode.executionStats}
\title{PlanNode.executionStats Object}
\usage{
PlanNode.executionStats()
}
\value{
PlanNode.executionStats object
}
\description{
PlanNode.executionStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The execution statistics associated with the node, contained in a group ofkey-value pairs. Only present if the plan was returned as a result of aprofile query. For example, number of executions, number of rows/time perexecution etc.
}
\seealso{
Other PlanNode functions: \code{\link{PlanNode.metadata}},
  \code{\link{PlanNode}}
}
