% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{PlanNode.metadata}
\alias{PlanNode.metadata}
\title{PlanNode.metadata Object}
\usage{
PlanNode.metadata()
}
\value{
PlanNode.metadata object
}
\description{
PlanNode.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attributes relevant to the node contained in a group of key-value pairs.For example, a Parameter Reference node could have the followinginformation in its metadata:    {      'parameter_reference': 'param1',      'parameter_type': 'array'    }
}
\seealso{
Other PlanNode functions: \code{\link{PlanNode.executionStats}},
  \code{\link{PlanNode}}
}
