% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{PlanNode}
\alias{PlanNode}
\title{PlanNode Object}
\usage{
PlanNode(PlanNode.metadata = NULL, PlanNode.executionStats = NULL,
  metadata = NULL, executionStats = NULL, shortRepresentation = NULL,
  index = NULL, displayName = NULL, childLinks = NULL)
}
\arguments{
\item{PlanNode.metadata}{The \link{PlanNode.metadata} object or list of objects}

\item{PlanNode.executionStats}{The \link{PlanNode.executionStats} object or list of objects}

\item{metadata}{Attributes relevant to the node contained in a group of key-value pairs}

\item{executionStats}{The execution statistics associated with the node, contained in a group of}

\item{shortRepresentation}{Condensed representation for SCALAR nodes}

\item{index}{The `PlanNode`'s index in node list}

\item{displayName}{The display name for the node}

\item{childLinks}{List of child node `index`es and their relationship to this parent}
}
\value{
PlanNode object
}
\description{
PlanNode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Node information for nodes appearing in a QueryPlan.plan_nodes.
}
\seealso{
Other PlanNode functions: \code{\link{PlanNode.executionStats}},
  \code{\link{PlanNode.metadata}}
}
