% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instanceConfigs.list}
\alias{projects.instanceConfigs.list}
\title{Lists the supported instance configurations for a given project.}
\usage{
projects.instanceConfigs.list(parent, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{If non-empty, `page_token` should contain a}

\item{pageSize}{Number of instance configurations to be returned in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}
}
