% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instances.create}
\alias{projects.instances.create}
\title{Creates an instance and begins preparing it to begin serving. Thereturned long-running operationcan be used to track the progress of preparing the newinstance. The instance name is assigned by the caller. If thenamed instance already exists, `CreateInstance` returns`ALREADY_EXISTS`.Immediately upon completion of this request:  * The instance is readable via the API, with all requested attributes    but no allocated resources. Its state is `CREATING`.Until completion of the returned operation:  * Cancelling the operation renders the instance immediately unreadable    via the API.  * The instance can be deleted.  * All other attempts to modify the instance are rejected.Upon completion of the returned operation:  * Billing for all successfully-allocated resources begins (some types    may have lower than the requested levels).  * Databases can be created in the instance.  * The instance's allocated resource levels are readable via the API.  * The instance's state becomes `READY`.The returned long-running operation willhave a name of the format `<instance_name>/operations/<operation_id>` andcan be used to track creation of the instance.  Themetadata field type isCreateInstanceMetadata.The response field type isInstance, if successful.}
\usage{
projects.instances.create(CreateInstanceRequest, parent)
}
\arguments{
\item{CreateInstanceRequest}{The \link{CreateInstanceRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}

Other CreateInstanceRequest functions: \code{\link{CreateInstanceRequest}}
}
