% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instances.getIamPolicy}
\alias{projects.instances.getIamPolicy}
\title{Gets the access control policy for an instance resource. Returns an emptypolicy if an instance exists but does not have a policy set.Authorization requires `spanner.instances.getIamPolicy` onresource.}
\usage{
projects.instances.getIamPolicy(GetIamPolicyRequest, resource)
}
\arguments{
\item{GetIamPolicyRequest}{The \link{GetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The Cloud Spanner resource for which the policy is being retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}

Other GetIamPolicyRequest functions: \code{\link{GetIamPolicyRequest}}
}
