% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instances.list}
\alias{projects.instances.list}
\title{Lists all instances in the given project.}
\usage{
projects.instances.list(parent, pageToken = NULL, pageSize = NULL,
  filter = NULL)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{If non-empty, `page_token` should contain a}

\item{pageSize}{Number of instances to be returned in the response}

\item{filter}{An expression for filtering the results of the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}
}
