% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instances.patch}
\alias{projects.instances.patch}
\title{Updates an instance, and begins allocating or releasing resourcesas requested. The returned long-runningoperation can be used to track theprogress of updating the instance. If the named instance does notexist, returns `NOT_FOUND`.Immediately upon completion of this request:  * For resource types for which a decrease in the instance's allocation    has been requested, billing is based on the newly-requested level.Until completion of the returned operation:  * Cancelling the operation sets its metadata's    cancel_time, and begins    restoring resources to their pre-request values. The operation    is guaranteed to succeed at undoing all resource changes,    after which point it terminates with a `CANCELLED` status.  * All other attempts to modify the instance are rejected.  * Reading the instance via the API continues to give the pre-request    resource levels.Upon completion of the returned operation:  * Billing begins for all successfully-allocated resources (some types    may have lower than the requested levels).  * All newly-reserved resources are available for serving the instance's    tables.  * The instance's new resource levels are readable via the API.The returned long-running operation willhave a name of the format `<instance_name>/operations/<operation_id>` andcan be used to track the instance modification.  Themetadata field type isUpdateInstanceMetadata.The response field type isInstance, if successful.Authorization requires `spanner.instances.update` permission onresource name.}
\usage{
projects.instances.patch(UpdateInstanceRequest, name)
}
\arguments{
\item{UpdateInstanceRequest}{The \link{UpdateInstanceRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}

Other UpdateInstanceRequest functions: \code{\link{UpdateInstanceRequest}}
}
