% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\name{projects.instances.testIamPermissions}
\alias{projects.instances.testIamPermissions}
\title{Returns permissions that the caller has on the specified instance resource.Attempting this RPC on a non-existent Cloud Spanner instance resource willresult in a NOT_FOUND error if the user has `spanner.instances.list`permission on the containing Google Cloud Project. Otherwise returns anempty set of permissions.}
\usage{
projects.instances.testIamPermissions(TestIamPermissionsRequest, resource)
}
\arguments{
\item{TestIamPermissionsRequest}{The \link{TestIamPermissionsRequest} object to pass to this method}

\item{resource}{REQUIRED: The Cloud Spanner resource for which permissions are being tested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/spanner/}{Google Documentation}

Other TestIamPermissionsRequest functions: \code{\link{TestIamPermissionsRequest}}
}
