% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{QueryPlan}
\alias{QueryPlan}
\title{QueryPlan Object}
\usage{
QueryPlan(planNodes = NULL)
}
\arguments{
\item{planNodes}{The nodes in the query plan}
}
\value{
QueryPlan object
}
\description{
QueryPlan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains an ordered list of nodes appearing in the query plan.
}
