% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ReadRequest}
\alias{ReadRequest}
\title{ReadRequest Object}
\usage{
ReadRequest(limit = NULL, index = NULL, keySet = NULL, columns = NULL,
  transaction = NULL, resumeToken = NULL, table = NULL)
}
\arguments{
\item{limit}{If greater than zero, only the first `limit` rows are yielded}

\item{index}{If non-empty, the name of an index on table}

\item{keySet}{Required}

\item{columns}{The columns of table to be returned for each row matching}

\item{transaction}{The transaction to use}

\item{resumeToken}{If this request is resuming a previously interrupted read,}

\item{table}{Required}
}
\value{
ReadRequest object
}
\description{
ReadRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Read andStreamingRead.
}
