% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ReadWrite}
\alias{ReadWrite}
\title{ReadWrite Object}
\usage{
ReadWrite()
}
\value{
ReadWrite object
}
\description{
ReadWrite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for read-write transactions.
}
