% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ResultSet}
\alias{ResultSet}
\title{ResultSet Object}
\usage{
ResultSet(metadata = NULL, stats = NULL, rows = NULL)
}
\arguments{
\item{metadata}{Metadata about the result set, such as row type information}

\item{stats}{Query plan and execution statistics for the query that produced this}

\item{rows}{Each element in `rows` is a row whose format is defined by}
}
\value{
ResultSet object
}
\description{
ResultSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Results from Read orExecuteSql.
}
