% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ResultSetMetadata}
\alias{ResultSetMetadata}
\title{ResultSetMetadata Object}
\usage{
ResultSetMetadata(rowType = NULL, transaction = NULL)
}
\arguments{
\item{rowType}{Indicates the field names and types for the rows in the result}

\item{transaction}{If the read or SQL query began a transaction as a side-effect, the}
}
\value{
ResultSetMetadata object
}
\description{
ResultSetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about a ResultSet or PartialResultSet.
}
