% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ResultSetStats.queryStats}
\alias{ResultSetStats.queryStats}
\title{ResultSetStats.queryStats Object}
\usage{
ResultSetStats.queryStats()
}
\value{
ResultSetStats.queryStats object
}
\description{
ResultSetStats.queryStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Aggregated statistics from the execution of the query. Only present whenthe query is profiled. For example, a query could return the statistics asfollows:    {      'rows_returned': '3',      'elapsed_time': '1.22 secs',      'cpu_time': '1.19 secs'    }
}
\seealso{
Other ResultSetStats functions: \code{\link{ResultSetStats}}
}
