% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ResultSetStats}
\alias{ResultSetStats}
\title{ResultSetStats Object}
\usage{
ResultSetStats(ResultSetStats.queryStats = NULL, queryStats = NULL,
  queryPlan = NULL)
}
\arguments{
\item{ResultSetStats.queryStats}{The \link{ResultSetStats.queryStats} object or list of objects}

\item{queryStats}{Aggregated statistics from the execution of the query}

\item{queryPlan}{QueryPlan for the query associated with this result}
}
\value{
ResultSetStats object
}
\description{
ResultSetStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional statistics about a ResultSet or PartialResultSet.
}
\seealso{
Other ResultSetStats functions: \code{\link{ResultSetStats.queryStats}}
}
