% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{RollbackRequest}
\alias{RollbackRequest}
\title{RollbackRequest Object}
\usage{
RollbackRequest(transactionId = NULL)
}
\arguments{
\item{transactionId}{Required}
}
\value{
RollbackRequest object
}
\description{
RollbackRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Rollback.
}
