% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Session}
\alias{Session}
\title{Session Object}
\usage{
Session(name = NULL)
}
\arguments{
\item{name}{Required}
}
\value{
Session object
}
\description{
Session Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A session in the Cloud Spanner API.
}
