% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{ShortRepresentation.subqueries}
\alias{ShortRepresentation.subqueries}
\title{ShortRepresentation.subqueries Object}
\usage{
ShortRepresentation.subqueries()
}
\value{
ShortRepresentation.subqueries object
}
\description{
ShortRepresentation.subqueries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A mapping of (subquery variable name) -> (subquery node id) for caseswhere the `description` string of this node references a `SCALAR`subquery contained in the expression subtree rooted at this node. Thereferenced `SCALAR` subquery may not necessarily be a direct child ofthis node.
}
\seealso{
Other ShortRepresentation functions: \code{\link{ShortRepresentation}}
}
