% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_functions.R
\docType{package}
\name{spanner_googleAuthR}
\alias{spanner_googleAuthR}
\alias{spanner_googleAuthR-package}
\title{Cloud Spanner API
Cloud Spanner is a managed, mission-critical, globally consistent and scalable relational database service.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:15:07
filename: /Users/mark/dev/R/autoGoogleAPI/googlespannerv1.auto/R/spanner_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
