% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{StructType}
\alias{StructType}
\title{StructType Object}
\usage{
StructType(fields = NULL)
}
\arguments{
\item{fields}{The list of fields that make up this struct}
}
\value{
StructType object
}
\description{
StructType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`StructType` defines the fields of a STRUCT type.
}
