% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Transaction}
\alias{Transaction}
\title{Transaction Object}
\usage{
Transaction(id = NULL, readTimestamp = NULL)
}
\arguments{
\item{id}{`id` may be used to identify the transaction in subsequent}

\item{readTimestamp}{For snapshot read-only transactions, the read timestamp chosen}
}
\value{
Transaction object
}
\description{
Transaction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A transaction.
}
