% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Type}
\alias{Type}
\title{Type Object}
\usage{
Type(structType = NULL, arrayElementType = NULL, code = NULL)
}
\arguments{
\item{structType}{If code == STRUCT, then `struct_type`}

\item{arrayElementType}{If code == ARRAY, then `array_element_type`}

\item{code}{Required}
}
\value{
Type object
}
\description{
Type Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Type` indicates the type of a Cloud Spanner value, as might be stored in atable cell or returned from an SQL query.
}
