% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{UpdateDatabaseDdlMetadata}
\alias{UpdateDatabaseDdlMetadata}
\title{UpdateDatabaseDdlMetadata Object}
\usage{
UpdateDatabaseDdlMetadata(statements = NULL, commitTimestamps = NULL,
  database = NULL)
}
\arguments{
\item{statements}{For an update this list contains all the statements}

\item{commitTimestamps}{Reports the commit timestamps of all statements that have}

\item{database}{The database being modified}
}
\value{
UpdateDatabaseDdlMetadata object
}
\description{
UpdateDatabaseDdlMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata type for the operation returned byUpdateDatabaseDdl.
}
