% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{UpdateDatabaseDdlRequest}
\alias{UpdateDatabaseDdlRequest}
\title{UpdateDatabaseDdlRequest Object}
\usage{
UpdateDatabaseDdlRequest(statements = NULL, operationId = NULL)
}
\arguments{
\item{statements}{DDL statements to be applied to the database}

\item{operationId}{If empty, the new update request is assigned an}
}
\value{
UpdateDatabaseDdlRequest object
}
\description{
UpdateDatabaseDdlRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enqueues the given DDL statements to be applied, in order but notnecessarily all at once, to the database schema at some point (orpoints) in the future. The server checks that the statementsare executable (syntactically valid, name tables that exist, etc.)before enqueueing them, but they may still fail uponlater execution (e.g., if a statement from another batch ofstatements is applied first and it conflicts in some way, or ifthere is some data-related problem like a `NULL` value in a column towhich `NOT NULL` would be added). If a statement fails, allsubsequent statements in the batch are automatically cancelled.Each batch of statements is assigned a name which can be used withthe Operations API to monitorprogress. See theoperation_id field for moredetails.
}
