% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{UpdateInstanceMetadata}
\alias{UpdateInstanceMetadata}
\title{UpdateInstanceMetadata Object}
\usage{
UpdateInstanceMetadata(instance = NULL, startTime = NULL,
  cancelTime = NULL, endTime = NULL)
}
\arguments{
\item{instance}{The desired end state of the update}

\item{startTime}{The time at which UpdateInstance}

\item{cancelTime}{The time at which this operation was cancelled}

\item{endTime}{The time at which this operation failed or was completed successfully}
}
\value{
UpdateInstanceMetadata object
}
\description{
UpdateInstanceMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata type for the operation returned byUpdateInstance.
}
