% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{UpdateInstanceRequest}
\alias{UpdateInstanceRequest}
\title{UpdateInstanceRequest Object}
\usage{
UpdateInstanceRequest(instance = NULL, fieldMask = NULL)
}
\arguments{
\item{instance}{Required}

\item{fieldMask}{Required}
}
\value{
UpdateInstanceRequest object
}
\description{
UpdateInstanceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for UpdateInstance.
}
\seealso{
Other UpdateInstanceRequest functions: \code{\link{projects.instances.patch}}
}
