% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_objects.R
\name{Write}
\alias{Write}
\title{Write Object}
\usage{
Write(columns = NULL, values = NULL, table = NULL)
}
\arguments{
\item{columns}{The names of the columns in table to be written}

\item{values}{The values to be written}

\item{table}{Required}
}
\value{
Write object
}
\description{
Write Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Arguments to insert, update, insert_or_update, andreplace operations.
}
